Trait pallet_anonymity_mining_claims::pallet::Config

pub trait Config<I: 'static = ()>: Config + Config + Config<I> {
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type PotId: Get<PalletId>;
    type APVanchorTreeId: Get<Self::TreeId>;
    type UnspentRootHistorySize: Get<Self::RootIndex>;
    type SpentRootHistorySize: Get<Self::RootIndex>;
    type Currency: MultiCurrency<Self::AccountId>;
    type VAnchor: VAnchorInterface<VAnchorConfigration<Self, I>> + VAnchorInspector<VAnchorConfigration<Self, I>>;
    type AnonymityPointsAssetId: Get<CurrencyIdOf<Self, I>>;
    type RewardAssetId: Get<CurrencyIdOf<Self, I>>;
    type NativeCurrencyId: Get<CurrencyIdOf<Self, I>>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
}

The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type PotId: Get<PalletId>

Account Identifier from which the internal Pot is generated.

type APVanchorTreeId: Get<Self::TreeId>

AP Vanchor Tree Id

type UnspentRootHistorySize: Get<Self::RootIndex>

History size of roots for each unspent tree

type SpentRootHistorySize: Get<Self::RootIndex>

History size of roots for each spent tree

type Currency: MultiCurrency<Self::AccountId>

Currency type for taking unspents

type VAnchor: VAnchorInterface<VAnchorConfigration<Self, I>> + VAnchorInspector<VAnchorConfigration<Self, I>>

VAnchor Interface

type AnonymityPointsAssetId: Get<CurrencyIdOf<Self, I>>

AP asset id

type RewardAssetId: Get<CurrencyIdOf<Self, I>>

Reward asset id

type NativeCurrencyId: Get<CurrencyIdOf<Self, I>>

Native currency id

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly reset parameters or otherwise alter privileged attributes.